+++ apps/endringsmelding-pleiepenger/src/app/api/fetchInitialData.ts
@@ -27,7 +27,7 @@ import søknadStateEndpoint, {
 } from './endpoints/søknadStateEndpoint';
 
 export const fetchInitialData = async (
-    tillattEndringsperiode: DateRange
+    tillattEndringsperiode: DateRange,
 ): Promise<{
     søker: Søker;
     k9saker: K9Sak[];
@@ -59,11 +59,13 @@ export const fetchInitialData = async (
                 k9saker = result.k9saker;
                 const periodeForArbeidsgiveroppslag = getPeriodeForArbeidsgiverOppslag(
                     result.dateRangeAlleSaker,
-                    tillattEndringsperiode
+                    tillattEndringsperiode,
                 );
                 if (!periodeForArbeidsgiveroppslag) {
                     return Promise.reject(
-                        getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode])
+                        getKanIkkeBrukeSøknadRejection([
+                            IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode,
+                        ]),
                     );
                 }
                 return arbeidsgivereEndpoint.fetch(periodeForArbeidsgiveroppslag);
@@ -103,7 +105,7 @@ export const fetchInitialData = async (
 
 const getKanIkkeBrukeSøknadRejection = (
     årsak: IngenTilgangÅrsak[],
-    ingenTilgangMeta?: IngenTilgangMeta
+    ingenTilgangMeta?: IngenTilgangMeta,
 ): Pick<SøknadInitialIkkeTilgang, 'årsak' | 'kanBrukeSøknad' | 'status' | 'ingenTilgangMeta'> => {
     return {
         status: RequestStatus.success,
@@ -116,14 +118,14 @@ const getKanIkkeBrukeSøknadRejection = (
 const kontrollerSaker = (
     k9sakerResult: K9SakResult[],
     antallSakerFørEndringsperiode: number,
-    tillattEndringsperiode: DateRange
+    tillattEndringsperiode: DateRange,
 ): Promise<{ k9saker: K9Sak[]; dateRangeAlleSaker: DateRange }> => {
     if (k9sakerResult.length === 0 && antallSakerFørEndringsperiode === 0) {
         return Promise.reject(getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.harIngenSak]));
     }
     if (k9sakerResult.length === 0 && antallSakerFørEndringsperiode > 0) {
         return Promise.reject(
-            getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode])
+            getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode]),
         );
     }
 
@@ -139,7 +141,7 @@ const kontrollerSaker = (
     }
     if (dateRangeUtils.dateRangesCollide([dateRangeAlleSaker, tillattEndringsperiode]) === false) {
         return Promise.reject(
-            getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode])
+            getKanIkkeBrukeSøknadRejection([IngenTilgangÅrsak.søknadsperioderUtenforTillattEndringsperiode]),
         );
     }
 
@@ -158,7 +160,7 @@ const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateR
                 JSON.stringify({
                     årsak: resultat.årsak,
                     sak: maskK9Sak(k9saker[0]),
-                })
+                }),
             );
         }
         if (k9saker.length > 1) {
@@ -166,7 +168,7 @@ const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateR
                 'IkkeTilgangSakInfo',
                 JSON.stringify({
                     årsak: resultat.årsak,
-                })
+                }),
             );
         }
     }
@@ -175,7 +177,7 @@ const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateR
 
 const hentOgKontrollerLagretSøknadState = async (
     søker: Søker,
-    k9saker: K9Sak[]
+    k9saker: K9Sak[],
 ): Promise<SøknadStatePersistence | undefined> => {
     const lagretSøknadState = await søknadStateEndpoint.fetch();
 
@@ -188,7 +190,7 @@ const hentOgKontrollerLagretSøknadState = async (
             søker,
             barnAktørId: lagretSøknadState.barnAktørId,
         },
-        k9saker
+        k9saker,
     );
     if (!isValid) {
         await søknadStateEndpoint.purge();