+++ apps/endringsmelding-pleiepenger/src/app/hooks/useSøknadInitialData.ts
@@ -50,7 +50,7 @@ const prepInitialData = (
         arbeidsgivere: Arbeidsgiver[];
         lagretSøknadState?: SøknadStatePersistence;
     },
-    tillattEndringsperiode: DateRange
+    tillattEndringsperiode: DateRange,
 ): SøknadInitialData => {
     const { arbeidsgivere, lagretSøknadState, k9saker, søker, antallSakerFørEndringsperiode } = loadedData;
 
@@ -68,8 +68,8 @@ const prepInitialData = (
                 appSentryLogger.logInfo(
                     'debug.maskedSakInfo',
                     JSON.stringify(
-                        getSakOgArbeidsgivereDebugInfo(k9saker[0], sak, arbeidsgivere, tillattEndringsperiode)
-                    )
+                        getSakOgArbeidsgivereDebugInfo(k9saker[0], sak, arbeidsgivere, tillattEndringsperiode),
+                    ),
                 );
             }
             return sak;