+++ apps/endringsmelding-pleiepenger/src/app/pages/velkommen/velkommenPageMessages.ts
@@ -11,9 +11,9 @@ export const velkommenPageMessages: MessageFileFormat = {
         'page.velkommen.guide.tittel': 'Hei {navn}',
         'page.velkommen.guide.ingress': 'Velkommen til søknad om ekstra omsorgsdager.',
         'page.velkommen.guide.tekst.1':
-            'Denne søknaden bruker du for å søke om ekstra omsorgsdager når du har barn med kronisk sykdom, funksjonshemming eller langvarig sykdom. Det er bare foreldre/fosterforeldre som barnet bor fast hos, som kan få ekstra omsorgsdager.',
+            'Denne søknaden bruker du for å søke om ekstra omsorgsdager når du har barn med kronisk sykdom, funksjonshemning eller langvarig sykdom. Det er bare foreldre/fosterforeldre som barnet bor fast hos, som kan få ekstra omsorgsdager.',
         'page.velkommen.guide.tekst.2':
-            'Foreldre må søke hver for seg. Hvis du har flere barn med kronisk sykdom, funksjonshemming eller langvarig sykdom, må du sende én søknad for hvert barn.',
+            'Foreldre må søke hver for seg. Hvis du har flere barn med kronisk sykdom, funksjonshemning eller langvarig sykdom, må du sende én søknad for hvert barn.',
 
         'page.velkommen.omSøknaden.tittel': 'Om utfylling av skjemaet',
         'page.velkommen.omSøknaden.1': 'Du får veiledning underveis om hva du skal fylle ut, og hvordan.',